[Feat] #10 - ProductRepositoryTest save 관련 테스트 작성

+++ baro-buyer/src/main/java/com/barofarm/buyer/product/application/dto/ProductDetailInfo.java
@@ -12,7 +12,7 @@ public record ProductDetailInfo(
     String productName,
     String description,
     ProductCategory productCategory,
-    Integer price,
+    Long price,
     Integer stockQuantity,
     ProductStatus productStatus,
     LocalDateTime createdAt,
